Use stdcall calling convention for platform ids
function called by icd loader on windows.

Without __stdcall there is a warning from debugger:
"Run-Time Check Failure #0 - The value of ESP was not properly saved across a
function call.  This is usually a result of calling a function declared with
one calling convention with a function pointer declared with a different calling
convention."